requirements.txt
pandas>=1.5.0
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=0.28.0
python-dotenv>=1.0.0
validators>=0.20.0
lxml>=4.9.0
